fix: move Firestore emulator credentials out of autoconfigure package
